Move button to compliance
