+++ apps/hyperdrive-trading/src/ui/hyperdrive/lp/RemoveLiquidityForm/RemoveLiquidityForm.tsx
@@ -12,13 +12,13 @@ import { calculateRatio } from "src/base/calculateRatio";
 import { calculateValueFromPrice } from "src/base/calculateValueFromPrice";
 import { isTestnetChain } from "src/chains/isTestnetChain";
 import { getHasEnoughBalance } from "src/token/getHasEnoughBalance";
-import { ConnectWalletButton } from "src/ui/base/components/ConnectWallet";
 import { LoadingButton } from "src/ui/base/components/LoadingButton";
 import { PrimaryStat } from "src/ui/base/components/PrimaryStat";
 import { formatBalance } from "src/ui/base/formatting/formatBalance";
 import { useActiveItem } from "src/ui/base/hooks/useActiveItem";
 import { useNumericInput } from "src/ui/base/hooks/useNumericInput";
 import { SwitchNetworksButton } from "src/ui/chains/SwitchChainButton/SwitchChainButton";
+import { ConnectWalletButton } from "src/ui/compliance/ConnectWallet";
 import { usePoolInfo } from "src/ui/hyperdrive/hooks/usePoolInfo";
 import { InvalidTransactionButton } from "src/ui/hyperdrive/InvalidTransactionButton";
 import { useLpSharesTotalSupply } from "src/ui/hyperdrive/lp/hooks/useLpSharesTotalSupply";
